pyproject.toml
[project]
name = "quarto-graft"
version = "0.0.2"
description = "Quarto GitHub Pages branch graft tool"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "Jinja2>=3.1",
  "packaging>=25.0",
  "pygit2>=1.19.0",
  "quarto-cli>=1.8.26",
  "questionary>=2.0.0",
  "ruamel-yaml>=0.18.16",
  "ruff>=0.14.8",
  "typer>=0.20.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
]

[project.scripts]
quarto-graft = "quarto_graft.cli:main"

[build-system]
requires = ["uv_build>=0.9.14,<0.10.0"]
build-backend = "uv_build"
